requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.9.0
groq>=0.11.0
jinja2>=3.1.2
python-multipart>=0.0.6
pypdf>=3.17.0
mammoth>=1.8.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
